requirements.txt
# Web scraping
beautifulsoup4>=4.12.0
playwright>=1.40.0
playwright-stealth>=1.0.6
requests>=2.31.0
lxml>=5.1.0

# Google Maps API
googlemaps>=4.10.0

# Data processing
pandas>=2.2.0

# Environment variables
python-dotenv>=1.0.0

# Geospatial calculations
geopy>=2.4.0

# HTTP requests with retry logic
tenacity>=8.2.0
